+++ engine/include/IOGPhysics.h
@@ -33,7 +33,7 @@ public:
 	virtual void RemoveObject (IOGPhysicalObject* _pObject) = 0;
 
 	// Update transforms.
-	virtual void Update (int _ElapsedTime) = 0;
+	virtual void Update (unsigned long _ElapsedTime) = 0;
 };
 
 #endif
